backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic-settings
numpy
pillow
tensorflow
